ch(project-structure): organisation [Finishes #168186623]
