+++ server/v1/test/session.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-undef */
 import { should,use,request } from 'chai';
 import chaiHttp from 'chai-http';
-import server from '../../index';
+import server from '../../../index';
